pyproject.toml
[project]
name = "pydna-mcp"
version = "0.1.0"
description = "MCP server for pydna - Python DNA simulation library"
readme = "README.md"
authors = [
    { name = "antonkulaga", email = "antonkulaga@gmail.com" }
]
requires-python = ">=3.10"
dependencies = [
    "eliot>=1.17.5",
    "pycomfort>=0.0.18",
    "fastmcp>=2.8.1",
    "fastapi>=0.115.13",
    "pydna>=5.2.0",
    "biopython>=1.84",
    "typer>=0.12.0",
    "scipy>=1.10.0",
    "numpy>=1.20.0",
    "matplotlib>=3.5.0",
    "requests>=2.25.0",
]

[project.scripts]
pydna-mcp = "pydna_mcp.server:cli_app_stdio"
server = "pydna_mcp.server:cli_app"
stdio = "pydna_mcp.server:cli_app_stdio"
sse = "pydna_mcp.server:cli_app_sse"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/pydna_mcp"]

[tool.hatch.build]
exclude = [
    "*.log",
    "logs/",
    ".venv/",
    ".pytest_cache/",
    "test/",
    "*.tmp",
    "temp*",
]

[dependency-groups]
dev = [
    "just-agents-core>=0.7.4",
    "pytest>=8.4.1",
    "pytest-asyncio>=0.26.0",
    "ruff>=0.12.0",
] 